src/cloud_function/cf_feedbacks/requirements.txt
Faker==37.1.0
google-cloud-secret-manager==2.23.3
google-cloud-bigquery==3.32.0